+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-* 2026-08-16 - AB#208: Add App Configuration to deployment.
+* 2026-08-18 - AB#208: Add App Configuration to AKS deployment.
+* 2026-08-16 - AB#208: Add App Configuration to ACA deployment.
 * 2026-08-14 - AB#177: Add AKS deployment workflow.
 * 2026-08-06 - AB#206: Validate certificate in the BSL.
 * 2026-08-03 - AB#206: Add certificate to api-to-bsl request